refactor(studio): modularize new page data and hooks
